+++ src/common/error.ts
@@ -1,0 +1,65 @@
+import * as vscode from 'vscode';
+import { Macro } from '../macro';
+import { openDocument } from './ui';
+
+export function showMacroErrorMessage(macro: Macro, error: Error | string): Promise<void> {
+  let message: string;
+  let selection: vscode.Range | undefined;
+  let stack: string | undefined;
+
+  if (typeof error === 'string') {
+    message = error;
+  } else {
+    message = error.message;
+    stack = error.stack;
+    if (error.stack) {
+      stack = error.stack;
+      selection = parseStack(stack);
+    }
+  }
+
+  return showErrorMessage(macro, message, stack, selection);
+
+  function parseStack(stack: string): vscode.Range | undefined {
+    const firstMatch = stack.match(/.+?(?<line>\d+)(:(?<offset>\d+))?$/m);
+    let position: vscode.Position | undefined;
+    if (firstMatch) {
+      const { line, offset } = firstMatch.groups!;
+      position = new vscode.Position(
+        parseInt(line) - 1,
+        offset ? (parseInt(offset) - 1) : 0);
+    }
+    return position && new vscode.Range(position, position);
+  }
+
+  async function showErrorMessage(macro: Macro, message: string, stack?: string, selection?: vscode.Range, modal = false): Promise<void> {
+    const actions = [
+      {
+        title: "Open",
+        execute: () => openDocument(macro.uri, { selection })
+      },
+      {
+        title: "Retry",
+        execute: () => vscode.commands.executeCommand('macros.run', macro.uri),
+      },
+    ];
+
+    const options: vscode.MessageOptions = {};
+    if (stack) {
+      if (modal) {
+        options.detail = stack;
+        options.modal = true;
+      } else {
+        actions.push({
+          title: "Details",
+          execute: () => showErrorMessage(macro, message, stack, selection, true),
+        });
+      }
+    }
+
+    const action = await vscode.window.showErrorMessage(message, options, ...actions);
+    if (action) {
+      await action.execute();
+    }
+  }
+}